html/semantics/embedded-content-0/the-iframe-element/iframe_sandbox_allow_script.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="uft-8">
    <title>HTML Test: iframe_sandbox_allow_scripts</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/embedded-content-0.html#attr-iframe-sandbox">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/embedded-content-0.html#the-iframe-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <iframe id="testIframe" src="support/sandbox_allow_script.html" sandbox="allow-same-origin" style="display:none"></iframe>

    <div id="log"></div>

    <script>
      var step1 = false;
      var t = async_test("iframe_sandbox_allow_scripts");

      setup({timeout:5000});
      window.addEventListener("message", callback1, false);

      function run(){
        window.removeEventListener("message", callback1, false);
        document.getElementById("testIframe").sandbox = "allow-scripts";
        document.getElementById("testIframe").contentWindow.location.reload();
        window.addEventListener("message", callback2, false);
      }

      function callback1(e) {
        step1= !step1;
      }

      function callback2(e) {
        t.step(function () {
          assert_false(step1, "Script", "[allow-scripts] is not set.");
          assert_equals(e.data, "Script executed", "[allow-scripts] is set.");
        });
        t.done();
      }

      setTimeout(run, 500);
    </script>
  </body>
</html>
